extract `arrayEquals` function to file and add dummy test for it
